requirement.txt
jax
mctx
numpy
matplotlib
seaborn
pygraphviz
chex
tqdm
pyyaml